move slot to its own file
