+++ packages/checkout/index.js
@@ -1,5 +1,6 @@
 export * from './totals';
 export * from './shipping';
+export * from './slot';
 export { default as ExperimentalOrderMeta } from './order-meta';
 export { default as ExperimentalOrderShippingPackages } from './order-shipping-packages';
 export { default as Panel } from './panel';
