+++ packages/checkout/order-meta/index.js
@@ -1,41 +1,21 @@
 /**
  * External dependencies
  */
-import { createSlotFill } from 'wordpress-components';
-import { Children, cloneElement } from '@wordpress/element';
 import classnames from 'classnames';
-import { CURRENT_USER_IS_ADMIN } from '@woocommerce/block-settings';
 import { useStoreCart } from '@woocommerce/base-hooks';
 
 /**
  * Internal dependencies
  */
-import BlockErrorBoundary from '../error-boundary';
+import { createSlotFill } from '../slot';
 
 const slotName = '__experimentalOrderMeta';
-const { Fill, Slot: OrderMetaSlot } = createSlotFill( slotName );
 
-function ExperimentalOrderMeta( { children } ) {
-	return (
-		<Fill>
-			{ ( fillProps ) => {
-				return Children.map( children, ( fill ) => {
-					return (
-						<BlockErrorBoundary
-							renderError={
-								CURRENT_USER_IS_ADMIN ? null : () => null
-							}
-						>
-							{ cloneElement( fill, fillProps ) }
-						</BlockErrorBoundary>
-					);
-				} );
-			} }
-		</Fill>
-	);
-}
+const { Fill: ExperimentalOrderMeta, Slot: OrderMetaSlot } = createSlotFill(
+	slotName
+);
 
-function Slot( { className } ) {
+const Slot = ( { className } ) => {
 	// We need to pluck out receiveCart.
 	// eslint-disable-next-line no-unused-vars
 	const { extensions, receiveCart, ...cart } = useStoreCart();
@@ -49,7 +29,7 @@ function Slot( { className } ) {
 			fillProps={ { extensions, cart } }
 		/>
 	);
-}
+};
 
 ExperimentalOrderMeta.Slot = Slot;
